Add XML data/format for SMP (#35)
* record truly random PRNG seeds

When the given seed is 0, then it makes a random seed
out of (for example) the microsecond time. The seed
to record is not the given value of 0, but the random
one it generated. This is necessary in order to correctly
reproduce the run.

* restored UTF8 characters (λ) that had been replaced by ASCII

* split BCN-stuff into separate file

* add SMP format in XML

Corrected dummyData_3Dim.csv to provide scenario name in "scenario name"
field, not PRNG seed.
Provide XML version of dummyData_3Dim, providing only the data in that
scenario. Also provide example.xml that illustrates two other blocks
of data. Finally, an XSD against which both example and dummyData_3Dim
validate, using XML Copy Editor