Store dimension names & correct/extend SQL logging (#37)
* dimension names now saved

There is now a table named DimensionDescription that stores the integer ID and a description name for each SMP dimension.  I have added a const to smp setting the max descrip to 256 chars

* progress on sql logging control

* Implement grouped table logging

This commit adds the infrastructure to control logging of data to SQLITE tables in logical groups. This is implemented both in the base kmodel class and extended in SMP, and extendable to all models. Wherever a logging function is called, it first checks for a specific group flag. Wherever logging is performed within a funnction, the group flag is also checked. There are currently 4 groups.