Removed unnecessary talking to console
In several places where data was logged to the SQLITE tables, there were unnecessary and virtually uninformative messages stating "Stored SQL for turn # of all estimators, actors, and positions". This was repeated many times in a block for each turn.  I commented out these lines of code.  It may be useful at some point in the future to turn it back on, but make the messages actually informative...