Correct Bargn table update
The code to update the Bargn table after resolution conditioned upon Turn_t, BargnId, Init_Act_i, and Recd_Act_j to get the correct row to update. However, when I reviewed the code, I observed it was not conditioning on ScenarioId, which means if multiple run with the same data were in the DB, they could potentially all be inadvertently updated.  I fixed this