Added Better Capabilities
Originally, I designed the capabilities vector for each actor to be strongly weighted for his sector and evenly across all others. I discovered, though, that the weight vectors are simply summed, so there is no useful concept in the model of "weight per sector".  I created a new set of capabilities vectors that are functions of the input relationship between sectors.  I also found a bug wherein the randomizer seed was used inside the LeonModel but not the base Model; obviously, I fixed this.