Update SOE Examples (#220)
As part of our recent GUI clarifications, the Scenario name & description default text (and soon to be tooltips) indicate that the fields can be used for either dataset name/desc or scenario name/desc (or a combination thereof).  I have updated the SOE example files Policy, Competitive, and Pol-Comp to reflect this.