Update Bargain Logging & Ref Runs (#227)
In conjunction with PR 226, this commit ensures minimal bargain data is logged into Bargn even with the --logmin flag on.  Note that the bargain entries are *not* updated from the queue resolution, by design.  Also, this commit updates the reference runs to reflect the changes from the same PR which prints the selected bargain ID at the end of bargain resolution.