get RPDemo working
RPDemo runs with minimal logging, without crashing and with correct results.
Changed exit(-1) to assert(false) in smpread, because it is easier to catch a
failed assert at the line it failed in the debugger, than to read an exit code and
try to figure out which of several lines might have set it.