Updated the external interface
A few changes made to existing Extern C interface function names. Biggest change was to simplify getVPHistory to just return a flat array of positions. The python demo script is updated to use this new function and parse the flat array into a 3d array of the most logical structure (states nested in dimensions nested in actors). Debug console printing is still on in both smp.cpp and pySMP.py